Moved tree cutting part and methods to BlockListener.
